Dockerfile combines python, node, mysql circleci images
When using circleci and multiple containers, all of the code we want to run is executed in the first container only which restricts our access to other containers. This means that if we have Python running in one container, then we cannot run commands with node/npm/mysql without manually installing each of these using run commands which is a real pain, especially when you have to configure them properly with correct permissions, etc. Therefore, if we combine the circleci images for python, node, and mysql, then we can run all of these in one container, simplifying the process.