Removed pymysql encoding for db
Currently, the build on circleci is failing because of an encoding error. Some absolutely brainless and horrendous person decided that mysql's utf8 would actually be some weird variant of utf8 that can only have 3 bytes and now we're stuck with their mess for eternity. This commit checks to see if pymysql's encoding arg is actually changing the db encoding or if it needs to be done manually. Trial and error programming for the win!