used actual code instead of method
Although the DRY principle exists, the code is much more readbale when placed separately in the two different files and arguably easier to debug with if anything goes wrong. Therefore. we'll ignore DRY for this case. Additionally, pytest was not running the test under TestAuthToken as it won't run tests that are classmethods. That particular test was converted to an instance method.