add iat to key & check jti for value
Only using the email address for the key can create some problems. For example, if the user logs in and logs out immediatelytwice in a row, then the token stored in the redis blacklist the first time the user logged out, will be replaced with the token that the user used the second time. This means that the first token becomes valid and can be used to authenticate. To avoid this problem, the iat is added to the key. Additionally, the blacklist loader not only checks if the key exists in the db, but also if the jti in the token is the same as the jti in the db. This makes such problems less likely.