used jti as the redis db key
jti is unique so it's better to make it the key for the db rather than a combination of the email address and the iat. The identity/email is stored as the value for the key in the db. This is because redis lists/sets don't allow an expiration date so you cannot automatically remove the tokens once they expire.