persist auth cookies & remove csrf cookie paths
Authentication cookies need to be persistent so they are not lost every time the user closes the window. Additionally, CSRF cookies were made globally accessible as the Javascript needs to access them prior to making the request for a new access token and if their paths are set to the api authentication endpoints, the client cannot attach them to the HTTP request.