cookie sharing between dev ports
In development, backend & frontend run on different ports. The same origin policy is followed & cookies can't be shared. This means that HTTP Only authentication cookies cannot be set on the client and accessed by the server. Some access control headers in development mode allows cookies to be shared between the ports.