made auth modals globally viewable
The authentication modals need to have the ability to get displayed from anywhere in the website. For example, when an unauthenticated user clicks a button to follow an author, the authentication modals need to be displayed immediately. Therefore, the authentication modal components need to be added to the `App.vue` component.