accept all json values for alterJsonKeys
JSON values can be null, boolean, number, string, array, or object. The previous implementation would throw a TypeError if a JSON object was not passed to it. However, keys of objects inside arrays can also be applied with the alteration function. The new implementation considers objects inside arrays and will return the same value passed to it if the value is not an array or object. This is a much more smoother, cleaner, and more maintainable implementation.